refactor(server): move the page layout widget grouping util to its own file
